change signature & exposed types
